+++ src/main/java/view/ErrorView.java
@@ -1,0 +1,7 @@
+package view;
+
+public class ErrorView {
+    public static void printErrorMessage(String message) {
+        System.out.println(message);
+    }
+}
