+++ src/main/java/view/ResultView.java
@@ -13,8 +13,4 @@ public class ResultView {
             System.out.println(ticket);
         }
     }
-
-    public static void printErrorMessage(String message) {
-        System.out.println(message);
-    }
 }
